pyproject.toml
[project]
name = "aether"
version = "0.1.0"
description = "Aether learning platform backend"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosmtplib==3.0.2",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asyncpg==0.30.0",
    "attrs>=25.4.0",
    "bcrypt==4.1.3",
    "blinker==1.9.0",
    "certifi>=2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer>=3.4.4",
    "click==8.3.0",
    "cryptography==46.0.2",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.117.1",
    "fastapi-mail==1.5.0",
    "fsrs>=6.3.0",
    "google-genai>=1.52.0",
    "greenlet==3.2.4",
    "h11>=0.16.0",
    "httpcore",
    "httptools==0.6.4",
    "httpx",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jupyter-client>=8.6.3",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.1.0",
    "notebook>=7.4.7",
    "pandas>=2.0.0",
    "pandas-stubs==2.3.2.250926",
    "passlib[bcrypt]>=1.7.4",
    "pinyin>=0.4.0",
    "pluggy==1.6.0",
    "psutil>=7.1.2",
    "psycopg==3.2.10",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pydantic==2.11.9",
    "pydantic-settings==2.11.0",
    "pydantic_core==2.33.2",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "pypinyin>=0.55.0",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-dotenv==0.5.2",
    "pytest-mock==3.15.1",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "resend>=2.5.0",
    "rsa==4.9.1",
    "setuptools==78.1.1",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.43",
    "starlette==0.48.0",
    "typing-inspection==0.4.1",
    "typing_extensions==4.15.0",
    "urllib3>=2.5.0",
    "uvicorn==0.37.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "wheel==0.45.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "black>=25.11.0",
]
